* tree-optimize.c (optimize_function_tree): Fix missed renamings.         * tree-ssa-pre.c (tree_perform_ssapre): Likewise.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/tree-ssa-20020619-branch@57316 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/function.c
@@ -3283,7 +3283,7 @@ insns_for_mem_hash (k)
 {
   /* Use the address of the key for the hash value.  */
   struct insns_for_mem_entry *m = (struct insns_for_mem_entry *) k;
-  return (hashval_t) m->key;
+  return (hashval_t)(size_t) m->key;
 }
 
 /* Return non-zero if K1 and K2 (two REGs) are the same.  */
